% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqmm-package.R
\docType{package}
\name{bqmm-package}
\alias{bqmm-package}
\title{bqmm: Bayesian Multilevel Quantile Regression}
\description{
Fits Bayesian mixed-effects (multilevel) quantile regression models using the
asymmetric Laplace working likelihood and Stan, with an \code{lme4}-style formula
interface, one or several quantiles per call, optional LKJ-correlated random
effects, post-hoc non-crossing rearrangement, and the Yang, Wang and He
(2016) posterior-variance correction for valid fixed-effect inference.
}
\references{
Yu, K. and Moyeed, R. A. (2001). Bayesian quantile regression.
\emph{Statistics & Probability Letters}, 54(3), 437-447.

Geraci, M. and Bottai, M. (2014). Linear quantile mixed models.
\emph{Statistics and Computing}, 24(3), 461-479.

Yang, Y., Wang, H. J. and He, X. (2016). Posterior inference in Bayesian
quantile regression with asymmetric Laplace likelihood.
\emph{International Statistical Review}, 84(3), 327-344.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvenkita/bqmm}
  \item \url{https://kvenkita.github.io/bqmm/}
  \item Report bugs at \url{https://github.com/kvenkita/bqmm/issues}
}

}
\author{
\strong{Maintainer}: Kailas Venkitasubramanian \email{kailasv@gmail.com} [copyright holder]

}
\keyword{internal}
