% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{bqmm_residuals_median}
\alias{bqmm_residuals_median}
\title{Residuals at the posterior-median fit (fixed + random part)}
\usage{
bqmm_residuals_median(object)
}
\description{
Residuals at the posterior-median fit (fixed + random part)
}
\keyword{internal}
