% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{check_convergence}
\alias{check_convergence}
\title{Warn on convergence problems}
\usage{
check_convergence(fit)
}
\arguments{
\item{fit}{A \code{stanfit} object.}
}
\description{
Warn on convergence problems
}
\keyword{internal}
