% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-core.R
\name{nobs.bqmm}
\alias{nobs.bqmm}
\title{Number of observations used in the fit}
\usage{
\method{nobs}{bqmm}(object, ...)
}
\arguments{
\item{object}{A \code{bqmm} fit.}

\item{...}{Unused.}
}
\value{
An integer, the number of observations.
}
\description{
Number of observations used in the fit
}
\examples{
\donttest{
fit <- bqmm(distance ~ age + (1 | Subject), data = nlme::Orthodont,
            tau = 0.5, chains = 1, iter = 300, refresh = 0, seed = 1)
nobs(fit)
}
}
