% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fixef}
\alias{ranef}
\alias{VarCorr}
\alias{as_draws}
\alias{posterior_predict}
\alias{posterior_epred}
\alias{log_lik}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lme4}{\code{\link[lme4]{fixef}}, \code{\link[lme4]{ranef}}, \code{\link[lme4]{VarCorr}}}

  \item{posterior}{\code{\link[posterior:draws]{as_draws}}}

  \item{rstantools}{\code{\link[rstantools]{log_lik}}, \code{\link[rstantools]{posterior_epred}}, \code{\link[rstantools]{posterior_predict}}}
}}

