% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{rho_tau}
\alias{rho_tau}
\title{Asymmetric Laplace check (pinball) loss}
\usage{
rho_tau(u, tau)
}
\arguments{
\item{u}{Numeric vector of residuals.}

\item{tau}{Quantile level in (0, 1).}
}
\value{
Numeric vector of loss values \code{u * (tau - (u < 0))}.
}
\description{
Asymmetric Laplace check (pinball) loss
}
\keyword{internal}
