% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-posterior.R
\name{tidy_param_names}
\alias{tidy_param_names}
\title{Map raw Stan parameter names to tidy bqmm names}
\usage{
tidy_param_names(nms, object)
}
\description{
Map raw Stan parameter names to tidy bqmm names
}
\keyword{internal}
